+++ project/assets_cartographer/manifests.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+"""
+Manifest registry for assets
+"""
+from parser import AutoAssetTagsManagerForKind
+
+# Default asset manifest
+manifest = AutoAssetTagsManagerForKind()