+++ project/settings_production.py
@@ -1,15 +1,14 @@
 from .settings import *
 
 # Update SITE infos to use the common port 80 to publish the webapp
-# Edit to another port if you dont want to serve the webapp on another port
 SITE_FIXED = {
     'name': "Recalbox Manager",
-    'ip': None, # If None find the ip automatically, else use a string to define another hostname
-    'port': None, # If None no port is appended to hostname, so the server have to be reachable from port 80
+    'ip': None, # If 'None' find the ip automatically. Use a string to define another ip/hostname
+    'port': None, # If 'None' no port is added to hostname, so the server have to be reachable from port 80
 }
 
 # Production path to the Recalbox logs file
 RECALBOX_LOGFILE_PATH = "/root/recalbox.log"
 
 # Use packaged assets
-ASSETS_COMPRESSED = True
+ASSETS_PACKAGED = True